chore: create `confirmabort` folder
Resolves: DCMAW-12237
