+++ features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/HandbackNavGraphProvider.kt
@@ -6,12 +6,12 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.squareup.anvil.annotations.ContributesMultibinding
-import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabortdesktopweb.ConfirmAbortDesktopWebViewModelModule
-import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabortdesktopweb.ConfirmAbortToDesktopWebScreen
-import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabortreturntodesktopweb.ConfirmAbortReturnToDesktopWebScreen
-import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabortreturntodesktopweb.ConfirmAbortReturnToDesktopWebViewModelModule
-import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmaborttomobileweb.ConfirmAbortToMobileWeb
-import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmaborttomobileweb.ConfirmAbortToMobileWebViewModelModule
+import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortdesktopweb.ConfirmAbortDesktopWebViewModelModule
+import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortdesktopweb.ConfirmAbortToDesktopWebScreen
+import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortreturntodesktopweb.ConfirmAbortReturnToDesktopWebScreen
+import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortreturntodesktopweb.ConfirmAbortReturnToDesktopWebViewModelModule
+import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmaborttomobileweb.ConfirmAbortToMobileWeb
+import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmaborttomobileweb.ConfirmAbortToMobileWebViewModelModule
 import uk.gov.onelogin.criorchestrator.features.handback.internal.navigatetomobileweb.WebNavigator
 import uk.gov.onelogin.criorchestrator.features.handback.internal.returntodesktopweb.ReturnToDesktopWebScreen
 import uk.gov.onelogin.criorchestrator.features.handback.internal.returntodesktopweb.ReturnToDesktopWebViewModelModule
