+++ features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/confirmabort/confirmabortdesktopweb/ConfirmAbortDesktopWebActions.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabortdesktopweb
+package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortdesktopweb
 
 sealed class ConfirmAbortDesktopWebActions {
     data object NavigateToReturnToDesktopWeb : ConfirmAbortDesktopWebActions()
