+++ features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/confirmabort/confirmabortdesktopweb/ConfirmAbortDesktopWebViewModel.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabortdesktopweb
+package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortdesktopweb
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
