+++ features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/confirmabort/confirmabortreturntodesktopweb/ConfirmAbortReturnToDesktopWebViewModelModule.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmaborttomobileweb
+package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortreturntodesktopweb
 
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.initializer
@@ -7,26 +7,21 @@ import com.squareup.anvil.annotations.ContributesTo
 import dagger.Module
 import dagger.Provides
 import uk.gov.onelogin.criorchestrator.features.handback.internal.analytics.HandbackAnalytics
-import uk.gov.onelogin.criorchestrator.features.session.internalapi.domain.SessionStore
 import uk.gov.onelogin.criorchestrator.libraries.di.CriOrchestratorScope
 import javax.inject.Named
 
 @Module
 @ContributesTo(CriOrchestratorScope::class)
-object ConfirmAbortToMobileWebViewModelModule {
-    const val FACTORY_NAME = "ConfirmAbortToMobileWebViewModelFactory"
+object ConfirmAbortReturnToDesktopWebViewModelModule {
+    const val FACTORY_NAME = "ConfirmAbortReturnToDesktopWebViewModelModuleFactory"
 
     @Provides
     @Named(FACTORY_NAME)
-    fun provideFactory(
-        sessionStore: SessionStore,
-        analytics: HandbackAnalytics,
-    ): ViewModelProvider.Factory =
+    fun provideViewModel(analytics: HandbackAnalytics): ViewModelProvider.Factory =
         viewModelFactory {
             initializer {
-                ConfirmAbortToMobileWebViewModel(
-                    sessionStore,
-                    analytics,
+                ConfirmAbortReturnToDesktopWebViewModel(
+                    analytics = analytics,
                 )
             }
         }
