+++ features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/confirmabort/confirmaborttomobileweb/ConfirmAbortToMobileWebConstants.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmaborttomobileweb
+package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmaborttomobileweb
 
 import androidx.annotation.StringRes
 import uk.gov.onelogin.criorchestrator.features.handback.internal.R
