+++ features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/confirmabort/confirmaborttomobileweb/ConfirmAbortToMobileWebViewModelModule.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabortdesktopweb
+package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmaborttomobileweb
 
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.initializer
@@ -7,21 +7,26 @@ import com.squareup.anvil.annotations.ContributesTo
 import dagger.Module
 import dagger.Provides
 import uk.gov.onelogin.criorchestrator.features.handback.internal.analytics.HandbackAnalytics
+import uk.gov.onelogin.criorchestrator.features.session.internalapi.domain.SessionStore
 import uk.gov.onelogin.criorchestrator.libraries.di.CriOrchestratorScope
 import javax.inject.Named
 
 @Module
 @ContributesTo(CriOrchestratorScope::class)
-object ConfirmAbortDesktopWebViewModelModule {
-    const val FACTORY_NAME = "ConfirmAbortToDesktopWebViewModelModuleFactory"
+object ConfirmAbortToMobileWebViewModelModule {
+    const val FACTORY_NAME = "ConfirmAbortToMobileWebViewModelFactory"
 
     @Provides
     @Named(FACTORY_NAME)
-    fun provideFactory(analytics: HandbackAnalytics): ViewModelProvider.Factory =
+    fun provideFactory(
+        sessionStore: SessionStore,
+        analytics: HandbackAnalytics,
+    ): ViewModelProvider.Factory =
         viewModelFactory {
             initializer {
-                ConfirmAbortDesktopWebViewModel(
-                    analytics = analytics,
+                ConfirmAbortToMobileWebViewModel(
+                    sessionStore,
+                    analytics,
                 )
             }
         }
